pyproject.toml
[tool.poetry]
name = "horssite-flask"
version = "0.1.0"
description = ""
authors = ["Tatiana_Garbuzova <tatiannna0697@gmail.com>"]
readme = "README.md"
packages = [{include = "horssite_flask"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
jinja2 = "^3.1.2"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
